+++ packages/akutil-common-test/src/index.js
@@ -1,4 +1,7 @@
-import afterMutations from './afterMutations';
 import waitUntil from './waitUntil';
+import afterMutations from './after-mutations';
+import hasClass from './hasClass';
+import { keydown, keyup } from './user-interaction';
+import getShadowRoot from './getShadowRoot';
 
-export { afterMutations, waitUntil };
+export { afterMutations, waitUntil, hasClass, getShadowRoot, keydown, keyup };