+++ packages/akutil-common-test/CHANGELOG.md
@@ -1,5 +1,15 @@
-<a name="1.2.0"></a>
-# 1.2.0 (2016-08-29)
+<a name="1.3.0"></a>
+# 1.3.0 (2016-08-31)
+
+
+
+<a name="akutil-common-test@1.3.0"></a>
+# akutil-common-test@1.3.0 (2016-08-30)
+
+
+### Features
+
+* **akutil-common-test:** Moved getShadowRoot and hasClass to akutil-common-test ([c246d64](https://bitbucket.org/atlassian/atlaskit/commits/c246d64))